+++ src/main/java/org/klang/torn/user/dto/FactionDTO.java
@@ -1,8 +1,7 @@
 package org.klang.torn.user.dto;
 
-import java.math.BigInteger;
 
 public record FactionDTO(
-        BigInteger money,
-        BigInteger points
+        Long money,
+        Long points
 ) {}
